improve css method and implement @css annotation support
